feat: changed current player in connect4 state to be jnp int8

+++ jumanji/environments/games/connect4/types.py
@@ -28,7 +28,7 @@ Board: TypeAlias = Array
 
 @dataclass
 class State:
-    current_player: int
+    current_player: jnp.int8
     board: Board
